change cast

+++ applications/RANSApplication/custom_processes/rans_nut_y_plus_wall_function_update_process.cpp
@@ -136,7 +136,8 @@ void RansNutYPlusWallFunctionUpdateProcess::Execute()
     VariableUtils().SetHistoricalVariableToZero(TURBULENT_VISCOSITY,
                                                 r_model_part.Nodes());
 
-    const double y_plus_limit = r_model_part.GetProcessInfo()[RANS_LINEAR_LOG_LAW_Y_PLUS_LIMIT];
+    const double y_plus_limit =
+        r_model_part.GetProcessInfo()[RANS_LINEAR_LOG_LAW_Y_PLUS_LIMIT];
 
     const int number_of_conditions = r_model_part.NumberOfConditions();
 #pragma omp parallel for
@@ -166,7 +167,7 @@ void RansNutYPlusWallFunctionUpdateProcess::Execute()
         NodeType& r_node = *(r_model_part.NodesBegin() + i_node);
         double& r_nut = r_node.FastGetSolutionStepValue(TURBULENT_VISCOSITY);
         const double number_of_neighbour_conditions =
-            static_cast<double>(r_node.GetValue(NUMBER_OF_NEIGHBOUR_CONDITIONS));
+            r_node.GetValue(NUMBER_OF_NEIGHBOUR_CONDITIONS);
         r_nut = RansCalculationUtilities::SoftMax(
             r_nut / number_of_neighbour_conditions, mMinValue);
         r_node.FastGetSolutionStepValue(VISCOSITY) =
